Fix Twitter Count #14
Twitter removed their counter from all share buttons. This change is based on the API from http://newsharecounts.com/

They count the new shares on their server so (for my site at least) the counter started from `0`, but I've tested it and it works fine.

They do mention:

> That's why we collected Twitter share counts for 1000000000 (yes, billion!) pages all around the Web before Twitter shut them down.  It's very likely that your share counts won't have to start from zero.

So for some sites it might not start from `0`. I did need to authorize their site for my account before though, I'm not sure if this is really needed or not, it may be that they do this by scraping the shares that are available on all the accounts that authenticated their site (but this is pure speculation).